Small changes before refactoty added
