+++ ImageFeed/Helpers/JSONDecoder+snakeCase.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+
 final class SnakeCaseJSONDecoder: JSONDecoder {
     override init() {
         super.init()
