+++ ImageFeed/Helpers/OAuth2TokenStorage.swift
