+++ ImageFeed/Models/ProfileModels/UserResult.swift
