+++ ImageFeed/Protocols/WebViewControllerDelegate.swift
@@ -5,8 +5,6 @@
 //  Created by Maksim Zakharov on 20.09.2024.
 //
 
-import Foundation
-
 protocol WebViewViewControllerDelegate: AnyObject {
     func webViewViewController(_ vc: WebViewViewController, didAuthenticateWithCode code: String)
     func webViewViewControllerDidCancel(_ vc: WebViewViewController)
