+++ ImageFeed/ViewConrollers+Navigation/TabBarController.swift
@@ -1,0 +1,12 @@
+//
+//  TabBarController.swift
+//  ImageFeed
+//
+//  Created by Maksim Zakharov on 13.10.2024.
+//
+
+import UIKit
+
+final class TabBarController: UITabBarController {
+    
+}
